messaging_api/requirements.txt
asgiref==3.8.1
Django==5.1.4
sqlparse==0.5.3
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
channels==4.2.0
channels_redis==4.2.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
kombu==5.4.2
msgpack==1.1.0
prompt_toolkit==3.0.48
PyJWT==2.10.1
python-dateutil==2.9.0.post0
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
